+++ src/surrounds.js
@@ -50,16 +50,6 @@ class Surrounds extends EventEmitter {
         this.sensorInputState = false;
     };
 
-    async startStopImpulseGenerator(state, timers = []) {
-        try {
-            //start impulse generator 
-            await this.zone.impulseGenerator.state(state, timers)
-            return true;
-        } catch (error) {
-            throw new Error(`Impulse generator start error: ${error}`);
-        }
-    }
-
     async prepareDataForAccessory() {
         try {
             //read dev info from file