+++ CHANGELOG.md
@@ -21,6 +21,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - For plugin < v6.1.0 use Homebridge UI <= v5.5.0
 - For plugin >= v6.1.0 use Homebridge UI >= v5.13.0
 
+## [6.1.7] - (15.02.2026)
+
+- refactor platform impulse generator
+- refactor data zones update
+- bump dependencies
+- cleanup
+
+## Changes
+
 ## [6.1.5] - (17.01.2026)
 
 ## Changes